requirements.txt
google-adk
gradio
mcp
httpx
python-dotenv
ddgs
beautifulsoup4
playwright
